Allow only logged in developer to update and delte himself

+++ src/db-actions/developers.js
@@ -1,6 +1,7 @@
 require('dotenv').config();
 
 const { Developer } = require('../models/Developer');
+const { PermissionError } = require('../validation/AuthErrors');
 
 const getDevelopers = async () =>
   await Developer.find({}).orFail(new Error('No documents found'));
@@ -15,8 +16,12 @@ const findDeveloperByEmail = async email => await Developer.findOne({ email });
 const createDeveloper = async developerData =>
   await new Developer(developerData).save();
 
-const updateDeveloper = async (id, updatedDeveloper) =>
-  await Developer.findByIdAndUpdate(
+const updateDeveloper = async (id, updatedDeveloper, loggedInDeveloper) => {
+  if (id !== loggedInDeveloper) {
+    const Error = PermissionError('Not permitted to update resource!');
+    throw new Error();
+  }
+  return await Developer.findByIdAndUpdate(
     id,
     { ...updatedDeveloper },
     {
@@ -24,13 +29,19 @@ const updateDeveloper = async (id, updatedDeveloper) =>
       useFindAndModify: false,
     },
   ).orFail(new Error(`Update failed, ${id} does not exist`));
+};
 
-const deleteDeveloper = async id =>
-  (await Developer.findByIdAndRemove(id, {
+const deleteDeveloper = async (id, loggedInDeveloper) => {
+  if (id !== loggedInDeveloper) {
+    const Error = PermissionError('Not permitted to delete resource!');
+    throw new Error();
+  }
+  return (await Developer.findByIdAndRemove(id, {
     useFindAndModify: false,
   }))
     ? id
     : new Error(`Delete failed, ${id} not found!`);
+};
 
 module.exports = {
   getDevelopers,
